Quick Start Guide

Introduction

Once Ipopt, PowerModelsDSSE and PowerModelsDistribution are installed, and a network data file (e.g. "case3_unbalanced.dss" in the package folder under ./test/data/extra/network) has been acquired as well as a measurement data file (e.g. "case3_unbalanced.csv" in the package folder under ./test/data/extra/measurement), an unbalanced AC Static State Estimation can be executed with,

using Ipopt
using PowerModelsDSSE
using PowerModelsDistribution

run_mc_se("case3_unbalanced.dss", "case3_unbalanced.csv", with_optimizer(Ipopt.Optimizer))

Network Data Input

Parsing OpenDSS files

To parse an OpenDSS file into PowerModelsDistribution's default ENGINEERING format, use the parse_file command:

eng = parse_file("case3_unbalanced.dss")

To get the MATHEMATICAL model it is possible to transform the data model using the transform_data_model command.

math = transform_data_model(eng)

Parsing ENWL files

To parse a specific feeder fdr of a network ntw of the ENWL data use:

data = parse_file(get_enwl_dss_path(ntw,fdr))

Parsing ENWL data requires the addition of a load profile before it can be used. This may be accomplished using

PowerModelsDSSE.insert_profiles! — Method

insert_profiles!(data, season, devices, pfs; t=missing, useactual=true)

This function adds the load profiles to the parsed ENWL ENGINEERING data data.

Arguments

season: "summer", "winter"

devices: "load", "pv", "ehp", "uchp"

pfs: power factor of the devices

t: time-step

Example

julia> data = parse_file(get_enwl_dss_path(10, 1))
julia> insert_profiles!(data, "summer", ["load", "pv"], [0.95, 0.90], t = 144)

Additionally, some functions are include specifically for the ENWL data to simplify the data in order to improve tractability.

PowerModelsDSSE.rm_enwl_transformer! — Method

rm_enwl_transformer!(data_eng)

This function removes the transformer from a parsed ENWL ENGINEERING data file.

PowerModelsDSSE.reduce_enwl_lines_eng! — Method

reduce_enwl_lines_eng!(data_eng)

This function removes all trailing lines from a parsed ENWL ENGINEERING data file.

PowerModelsDSSE.reduce_enwl_lines_math! — Method

reduce_enwl_lines_math!(data_math)

This function removes all trailing lines from a parsed ENWL MATHEMATICAL data file.

Measurement Data Input

Adding the measurements to a MATHEMATICAL data dictionary may be accomplished through:

PowerModelsDSSE.add_measurements! — Method

add_measurements!(data::Dict, meas_file::String; actual_meas::Bool = false, seed::Int=0)

Add measurement data from separate CSV file to the PowerModelsDistribution data dictionary.

Arguments

data: data dictionary in a format usable with PowerModelsDistribution

meas_file: path to and name of file with measurement data

actual_meas: default is false.

false: the "val" in meas_file are not actual measurements, e.g., error-free powerflow results. In this case, a fake measurement is built, extracting a value from the given distribution.

true: the "val" column in meas_file are actual measurement values, i.e., with an error. No other processing is required.

seed: random seed value to make the results reproducible. It is an argument and not a fixed value inside the function to allow, e.g., the exploration of different MonteCarlo scenarios

Furthermore, functionality is included to write a measurement file, i.e., a csv-file based on powerflow results, using:

PowerModelsDSSE.write_measurements! — Method

write_measurements!(model::Type, data::Dict, pf_results::Dict, path::String)

Function to write a measurement file, i.e., a csv-file, to a specific path based on the power flow.