Mathematical Model of the State Estimation Criteria

Let X be the random variable associated to a measurement m ∈ 𝓜 and x ∈ 𝓧 the related variable, where:

𝓜 denotes the set of measurements,

𝓧 denotes the (extended) variable space of the OPF problem.

The state of a power system can be determined based on a specific estimation criterion. The state estimator criteria can be classified into two groups based on the random variable X:

Gaussian

wlav: weighted least absolute value (exact)

rwlav: relaxed weighted least absolute value (exact relaxation)

wls: weighted least square (exact)

rwls: relaxed weighted least square (exact relaxation)

Non-Gaussian

mle: maximum likelihood estimation (exact)

The user has to specify the criterion for each measurement individually. However, if all measurements can be described by the same criterion, it is sufficient to set the criterion name in the se_settings (Input Data Format), as PMDSE will automatically call assign_unique_individual_criterion!().

PowerModelsDistributionStateEstimation.assign_unique_individual_criterion! — Method

assign_unique_individual_criterion!(data::Dict)
- data: `MATHEMATICAL` data model of the network
Assigns the criterion in data["se_settings"]["criterion"] to all individual measurements.

To use a combination of different criteria, the user should provide information in data["meas"], under a crit key. For example: data["meas"]["1"]["crit"] = "rwlav" and data["meas"]["2"]["crit"] = "mle". A basic helper function to assign different criteria to different measurement is provided:

PowerModelsDistributionStateEstimation.assign_basic_individual_criteria! — Method

assign_basic_individual_criteria!(data::Dict; chosen_criterion::String="rwlav")

Basic function that assigns individual criteria to measurements in data["meas"]. For each measurement, if the distribution type of at least one of the phases is normal, the criterion defaults to the chosen_criterion. Otherwise, it is assigned the 'mle' criterion. The function takes as input either a single measurement dictionary, e.g., data["meas"]["1"] or the full MATHEMATICAL data model.

Furthermore, a rescaler can be introduced to improve the convergence of the state estimation. The user has to specify the rescaler through the se_settings (Input Data Format). If no rescaler is specified, it will default to 1.0.

Gaussian State Estimation Criteria

WLAV and rWLAV

The WLAV criterion represents the absolute value norm (p=1) and is given by

\[\begin{eqnarray} \rho_{m} &= \frac{| x - \mu_{m} |}{\text{rsc} \cdot \sigma_{m}},\quad m \in \mathcal{M}: m \to x \in \mathcal{X}, \end{eqnarray}\]

where:

ρ denotes the residual associated with a measurement $m$,

x denotes the variable corresponding to a measurement $m$.

μ denotes the measured value, i.e., expectation 𝐄(X),

σ denotes the the measurement error, i.e., standard deviation √(𝐕(X)),

rsc denotes the rescaler.

Solving a state estimation using the WLAV criterion is non-trivial as the absolute value function is not continuously differentiable. This drawback is lifted by its exact linear relaxation: rWLAV[1]. The rWLAV criterion is given by

\[\begin{eqnarray} \rho_{m} &\geq \frac{ x_{m} - \mu_{m} }{\text{rsc} \cdot \sigma_{m}},\quad m \in \mathcal{M}: m \to x_{m} \in \mathcal{X}, \\ \rho_{m} &\geq - \frac{ x_{m} - \mu_{m} }{\text{rsc} \cdot \sigma_{m}},\quad m \in \mathcal{M}: m \to x_{m} \in \mathcal{X}, \\ \end{eqnarray}\]

WLS and rWLS

The WLS criterion represents the Eucledian norm (p=2) and is given by

\[\begin{eqnarray} \rho_{m} &= \frac{( x_{m} - \mu_{m} )^{2}}{\text{rsc} \cdot \sigma_{m}^{2}},\quad m \in \mathcal{M}: m_{m} \to x \in \mathcal{X}. \end{eqnarray}\]

The rWLS criterion relaxes the former as a cone and is given by

\[\begin{eqnarray} rsc \cdot \sigma_{m}^{2} \cdot \rho_{m} &\geq ( x_{m} - \mu_{m} )^{2},\quad m \in \mathcal{M}: m \to x_{m} \in \mathcal{X}. \end{eqnarray}\]

Non-Gaussian State Estimation Criteria

Maximum Likelihood Estimation

The maximum likelihood criterion links the measurement residual to the logpdf of the associated distribution and is given by

\[\begin{eqnarray} \rho_{m} &= - \text{rsc} \cdot \text{logpdf}_{m}(x) + \text{shf},\quad m \in \mathcal{M}: m \to x \in \mathcal{X}. \end{eqnarray}\]

where shf denotes a shift setting the minimum value of the residual to zero.

To avoid the use of automatic differentiation, the first derivative (gradlogpdf) is provided by Distributions.jl and the second derivative (heslogpdf) is provided internally.

Supported Distributions

Currently, the following univariate continuous distributions are supported through the Distributions.jl package:

Exponential

Weibull

Normal

Log-Normal

Gamma

Beta

Extended Beta

PowerModelsDistributionStateEstimation.ExtendedBeta — Type

ExtendedBeta

The extended beta distribution with shape parameters α and β, and optional support parameters min and max has a probability density function

\[f(x, α, β, \text{min}, \text{max}) = \begin{cases} 0, &\text{if:}~x < \text{min}, \\ \frac{(x-\text{min})^{α-1} (\text{max}-x)^{β-1}}{B(α,β) (\text{max}-\text{min})^{α+β-1}}, &\text{if:}~\text{min} ≤ x ≤ \text{max}, \\ 0, &\text{if:}~x > max, \end{cases}\]

where B(α,β) is a Beta function.

Furthermore, Gaussian Mixture Models (GMM) can also be used in the maximum likelihood estimation. These can be generated via GaussianMixtures.jl, and PMDSE provides logpdf,gradlogpdf and heslogpdf of the resulting GMM. Using GMM are particularly convenient to model non-parametric distributions.

TODO: add docs on Gaussian Mixture Model

The user has to specify the number_of_gaussian through the se_settings (Input Data Format). If no number is specified, it will default to 10.

1Note that this relaxation is only exact in the context of minimization problem.